requirements.txt
pyyaml
pathlib2
